clients/web: use dedicated library to connect to SSE
The builtin EventSource class doesn't support to pass custom headers to the request, like Authorization.

From what I read, this class is somewhat deprecated and browser developers encourage to use `fetch` instead. The thing is, it's a much lower level API which doesn't do the work of reading and parsing the stream. Ref: https://github.com/whatwg/html/issues/2177

I found `event-source-plus`, a library which does all that work on top of `fetch`.